Move Enum class to enum module.
